logg leverandør hvis vi har leverandør og konsumer eller "MANGLER" hvis ikke
