+++ src/main/kotlin/no/nav/familie/ks/barnehagelister/rest/DefaultBarnehagelisterController.kt
@@ -27,8 +27,12 @@ class DefaultBarnehagelisterController(
 
     private fun validerGodkjentLeverandør(request: HttpServletRequest) {
         val supplierId = request.hentSupplierId()
+        if (supplierId == null) {
+            throw UkjentLeverandørFeil("No supplier in request.")
+        }
+
         if (supplierId !in godkjenteLeverandører.leverandorer.map { it.orgno }) {
-            throw UkjentLeverandørFeil("Leverandør med orgnr ${supplierId.substringAfter(":")} er ikke en godkjent leverandør")
+            throw UkjentLeverandørFeil("Supplier with orgno ${supplierId.substringAfter(":")} is not a known supplier.")
         }
     }
